+++ rodjlib.py
@@ -744,26 +744,3 @@ def spawn_rodbezj_element(elem):
         return{'FINISHED'}
 # -----------------------------------------------------------
 # end of spawn_rodbezj_element(elem) function
-
-## Bevel object panel
-class Data_OT_MBDyn_Elems_Rodj(bpy.types.Panel):
-    """ Beam operators panel in data properties panel """
-    bl_label = "MBDyn rod joint props"
-    bl_space_type = 'PROPERTIES'
-    bl_region_type = 'WINDOW'
-    bl_context = 'data'
-
-    def draw(self, context):
-        ed = context.scene.mbdyn.elems
-        curve = context.curve
-        layout = self.layout
-        col = layout.column()
-        
-        try:
-            if ed[context.object.name].type == 'rodj':
-                col.operator(Object_OT_MBDyn_load_section.bl_idname, text="Load profile (Selig)")
-                col.operator(Object_OT_MBDyn_update_beam3.bl_idname, text="Update configuration")
-        except KeyError:
-            pass
-# -----------------------------------------------------------
-# end of Data_OT_MBDyn_Elems_Beams class
